requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.34.0
django-appconf==1.0.2
django-crispy-forms==1.7.0
django-extensions==2.0.7
django-imagekit==4.0.2
django-widget-tweaks==1.4.1
idna==2.6
oauthlib==2.0.6
pilkit==2.0
Pillow==5.1.0
psycopg2==2.7.3.2
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
shortuuid==0.5.0
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
